+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/SVGComponents/ZoomEventLayer/ZoomEventLayer.js
@@ -1,33 +1,52 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import styled, { css } from 'styled-components'
+import styled, { css, withTheme } from 'styled-components'
+import withKeyZoom from '../../../../withKeyZoom'
 
 const StyledRect = styled.rect`
   ${props => props.panning ?
    css`cursor: move;` :
    css`cursor: inherit;`}
   overscroll-behavior: none;
+
+  &:focus {
+    ${props => 
+      css`
+        outline-color: ${props.focusColor};
+        border: solid thick ${props.focusColor};
+        box-shadow: 0 0 4px 4px ${props.focusColor};
+      `
+    }
+  }
 `
 
 function ZoomEventLayer (props) {
   const {
-    onDoubleClick = () => {},
+    height,
+    left,
+    onDoubleClick,
+    onKeyDown,
     onMouseDown,
     onMouseEnter,
     onMouseMove,
     onMouseUp,
     onMouseLeave,
-    onWheel = () => {},
-    panning = false,
-    parentHeight,
-    parentWidth
+    onWheel,
+    panning,
+    theme,
+    top,
+    width,
+    ...rest
   } = props
 
+  const focusColor = theme.global.colors[theme.global.colors.focus]
   return (
     <StyledRect
-      height={parentHeight}
       fill='transparent'
+      focusColor={focusColor}
+      height={height}
       onDoubleClick={onDoubleClick}
+      onKeyDown={onKeyDown}
       onMouseDown={onMouseDown}
       onMouseEnter={onMouseEnter}
       onMouseMove={onMouseMove}
@@ -35,13 +54,33 @@ function ZoomEventLayer (props) {
       onMouseLeave={onMouseLeave}
       onWheel={onWheel}
       panning={(panning) ? 'true' : undefined}
-      width={parentWidth}
+      transform={`translate(${left}, ${top})`}
+      width={width}
+      {...rest}
     />
   )
 }
 
+ZoomEventLayer.defaultProps = {
+  left: 0,
+  onDoubleClick: () => {},
+  onKeyDown: () => {},
+  onMouseEnter: () => {},
+  onWheel: () => {},
+  panning: false,
+  theme: {
+    global: {
+      colors: {}
+    }
+  },
+  top: 0
+}
+
 ZoomEventLayer.propTypes = {
+  height: PropTypes.number.isRequired,
+  left: PropTypes.number,
   onDoubleClick: PropTypes.func,
+  onKeyDown: PropTypes.func,
   onMouseDown: PropTypes.func.isRequired,
   onMouseEnter: PropTypes.func,
   onMouseMove: PropTypes.func.isRequired,
@@ -49,8 +88,10 @@ ZoomEventLayer.propTypes = {
   onMouseLeave: PropTypes.func.isRequired,
   onWheel: PropTypes.func,
   panning: PropTypes.bool,
-  parentHeight: PropTypes.number.isRequired,
-  parentWidth: PropTypes.number.isRequired
+  theme: PropTypes.object,
+  top: PropTypes.number,
+  width: PropTypes.number.isRequired
 }
 
-export default ZoomEventLayer
+export default withTheme(ZoomEventLayer)
+export { ZoomEventLayer }